Change stat variables to unsigned long longs

+++ Lab 3/SkipList.h
@@ -45,6 +45,6 @@ private:
 
 	int height;
 	int amountOfItems;
-	unsigned int numberOfKeyComparisonsMade;
-	unsigned int numberOfReferenceChanges;
+	unsigned long long numberOfKeyComparisonsMade;
+	unsigned long long numberOfReferenceChanges;
 };
